Add gMSA password fetch and decode
gMSA password is fetched using LDAP query over kerberos.
For kerberos to work, krb ticket from a principal in the
PrincipalsAllowedToRetrieveManagedPassword list for the gMSA account
is required as a pre-requisite.
ldapsearch and kinit utilities are also required as pre-requisites.

Added decoding of gMSA password using blob structure at:
https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-adts/a9019740-3d73-46ef-a9ae-3ea8eb86ac2

The decoded password can be used to do kinit and get a new krb ticket.

TBD: krb5-cc-name should be changed so that the machine krb ticket is
not overwritten.