This is a defensive change, it restarts ecs if credentials-fetcher gets stuck. This is not for any known issue, this is for unknown 'heisenbugs'.
Added an optional script to the systemd service file of credentials-fetcher. This script gets executed 'systemctl restart credentials-fetcher'.
The restart of ecs occurs only once, the human operator has to delete the lock file to allow the next restart.

The script is optional due to the '-' at the front, so, even if it failed, it will not affect the restart of credentials-fetcher.